refactor: adjustments made to `csproj` files and code namespaces
This commit includes vast changes across several `csproj` files and code files due to the refactor of the project. The 'RunnerTest' references have been removed and replaced with references to the 'Kraken.REST.API' project in multiple files. The namespaces of several classes have also been updated from 'RunnerTest' to 'Kraken.REST.API'. The purpose of this refactor was to ensure the project is properly structured and code dependencies are correctly managed. The 'RunnerTest' project files and associated references were determined to be unnecessary, so they have been removed to streamline the project structure.